[FIX] Room position improvements
- Do not reset character position in some cases
- Better initial position
